Change output of the script to make communication with axisem easier.
Now this script generates a list of paths with the sph files generated with the project. Each of these paths is paired with a name for the axisem run. The name is like: 'StagyyName'-'ThermodynamicDatasetInfo'-'Gyr'-'isnap(Stagpy)'. The list is saved as {ProjectName}-runList. This list can be read by the run_manager.sh provided. The submit.csh must also be the one provided.