Mirror AxiSEM 'half Earth' grid to get a complete circle
This is necessary because source position in AxiSEM is arbitrary. This means that source, receivers and heterogeneity positions (latitude/longitude) will be rotated so that the source coincides with the axis (pole). Your velocity map will then need to span the whole 2pi domain, otherwise a piece will be missing after rotation. In other words, the interpolation must be carried out over a circular (as opposed to a half-circular) domain.