Optimize KDTree creation and query
You only need to create a tree for each thermodynamic dataset, because the tree only depends on the P,T range of the thermodynamic dataset, assuming this range is the same for each composition. These datasets might be different for each model or not. We optimize the case when they are all the same. The tree must be queried for each time step, but only once for each composition, since the query only depends by the the P, T fields of the model (which change in time, but are all the same in the stagyy domain).