Handle resampling on a rectangular grid
If the grid is rectangular (which will never be the case for an axisem mesh), the interpolate method can handle both downsampling and upsampling. In either case, the original P, T, C fields are interpolated on the new mesh: however, in the former case, a simple low-pass filter is first applied against aliasing. CANNOT HANDLE non-uniform sampling along the theta axis!